requirements.txt
streamlit>=1.30.0
pillow>=10.0.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
python-dateutil>=2.8.2
